+++ src/main/java/uk/gov/hmcts/reform/prl/controllers/ReturnApplicationReturnMessageController.java
@@ -24,7 +24,6 @@ import uk.gov.hmcts.reform.prl.handlers.CaseEventHandler;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.services.AuthorisationService;
 import uk.gov.hmcts.reform.prl.services.EventService;
-import uk.gov.hmcts.reform.prl.services.MiamPolicyUpgradeFileUploadService;
 import uk.gov.hmcts.reform.prl.services.ReturnApplicationService;
 import uk.gov.hmcts.reform.prl.services.UserService;
 import uk.gov.hmcts.reform.prl.services.tab.alltabs.AllTabServiceImpl;
@@ -33,8 +32,6 @@ import uk.gov.hmcts.reform.prl.utils.CaseUtils;
 import java.util.Map;
 
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
-import static org.apache.commons.lang3.ObjectUtils.isNotEmpty;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.C100_CASE_TYPE;
 import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.INVALID_CLIENT;
 
 @RestController
@@ -46,7 +43,6 @@ public class ReturnApplicationReturnMessageController extends AbstractCallbackCo
     private final AuthorisationService authorisationService;
     private final CaseEventHandler caseEventHandler;
 
-    private final MiamPolicyUpgradeFileUploadService miamPolicyUpgradeFileUploadService;
 
     @Autowired
     public ReturnApplicationReturnMessageController(ObjectMapper objectMapper,
@@ -55,15 +51,13 @@ public class ReturnApplicationReturnMessageController extends AbstractCallbackCo
                                                     ReturnApplicationService returnApplicationService,
                                                     AllTabServiceImpl allTabsService,
                                                     AuthorisationService authorisationService,
-                                                    CaseEventHandler caseEventHandler,
-                                                    MiamPolicyUpgradeFileUploadService miamPolicyUpgradeFileUploadService) {
+                                                    CaseEventHandler caseEventHandler) {
         super(objectMapper, eventPublisher);
         this.userService = userService;
         this.returnApplicationService = returnApplicationService;
         this.allTabsService = allTabsService;
         this.authorisationService = authorisationService;
         this.caseEventHandler = caseEventHandler;
-        this.miamPolicyUpgradeFileUploadService = miamPolicyUpgradeFileUploadService;
     }
 
     @PostMapping(path = "/return-application-return-message", consumes = APPLICATION_JSON, produces = APPLICATION_JSON)
@@ -110,14 +104,7 @@ public class ReturnApplicationReturnMessageController extends AbstractCallbackCo
             // Refreshing the page in the same event. Hence no external event call needed.
             // Getting the tab fields and add it to the casedetails..
             Map<String, Object> caseDataUpdated = callbackRequest.getCaseDetails().getData();
-            if (C100_CASE_TYPE.equals(CaseUtils.getCaseTypeOfApplication(caseData))
-                && isNotEmpty(caseData.getMiamPolicyUpgradeDetails())) {
-                caseData = miamPolicyUpgradeFileUploadService.renameMiamPolicyUpgradeDocumentWithoutConfidential(
-                    caseData
-                );
-                allTabsService.getNewMiamPolicyUpgradeDocumentMap(caseData, caseDataUpdated);
-
-            }
+            caseData = returnApplicationService.updateMiamPolicyUpgradeDataForConfidentialDocument(caseData, caseDataUpdated);
             caseDataUpdated.put("taskListReturn", returnApplicationService.getReturnMessageForTaskList(caseData));
 
             String updatedTaskList = caseEventHandler.getUpdatedTaskList(caseData);
