+++ src/main/java/uk/gov/hmcts/reform/prl/services/ReturnApplicationService.java
@@ -11,17 +11,25 @@ import uk.gov.hmcts.reform.prl.enums.RejectReasonEnum;
 import uk.gov.hmcts.reform.prl.models.Element;
 import uk.gov.hmcts.reform.prl.models.complextypes.PartyDetails;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
+import uk.gov.hmcts.reform.prl.services.tab.alltabs.AllTabServiceImpl;
+import uk.gov.hmcts.reform.prl.utils.CaseUtils;
 
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import static java.util.Optional.ofNullable;
+import static org.apache.commons.lang3.ObjectUtils.isNotEmpty;
+import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.C100_CASE_TYPE;
 import static uk.gov.hmcts.reform.prl.services.validators.EventCheckerHelper.allNonEmpty;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class ReturnApplicationService {
+
+    private final MiamPolicyUpgradeFileUploadService miamPolicyUpgradeFileUploadService;
+    private final AllTabServiceImpl allTabsService;
 
     public String getLegalFullName(CaseData caseData) {
 
@@ -139,5 +147,17 @@ public class ReturnApplicationService {
 
         return returnMsgStr.toString();
 
+    }
+
+    public CaseData updateMiamPolicyUpgradeDataForConfidentialDocument(CaseData caseData, Map<String, Object> caseDataUpdated) {
+        if (C100_CASE_TYPE.equals(CaseUtils.getCaseTypeOfApplication(caseData))
+            && isNotEmpty(caseData.getMiamPolicyUpgradeDetails())) {
+            caseData = miamPolicyUpgradeFileUploadService.renameMiamPolicyUpgradeDocumentWithoutConfidential(
+                caseData
+            );
+            allTabsService.getNewMiamPolicyUpgradeDocumentMap(caseData, caseDataUpdated);
+
+        }
+        return caseData;
     }
 }
